Move matrix product Ops to sub-module
